requirements.txt
python-dotenv
setuptools
llama-index
matplotlib
Pillow
streamlit